tests(python): add tests for python binding
You can add ``.py`` tests for python binding at ``tests/python/``. For instance,
we have ``tests/python/is_var.py``:

```
import lean
x = lean.mk_var(0)
print(lean.is_var(x))
```

It uses the return value of python interpreter to determine whether a
test is passed or not. In addition to that, we check whether the result
output contains "swig/python detected a memory leak of type" or not. If
so, we report a memory-leak.

Python tests have the label ``python`` which can be used by ctest:

```
ctest -L python
```