feat(CMakeLists.txt): include cpp14_lang/sized_deallocation.cmake
Previously, it downloaded cpp14_lang/sized_deallocation.cmake file from
Github during cmake was running. This worked in most cases but failed in an
environment where network connection was not provided. A particular
example is to build Ubuntu packages. See [1] for details.

This patch is to include cpp14_lang/sized_deallocation.cmake file in the
repository. It also set COMP_CMAKE_PATH to indicate the location of the
pre-downloaded file. 'CACHE INTERNAL ""' is required to set the scope of
COMP_CMAKE_PATH file so that it's visible inside of
'CompBase.cmake'. See [2] for details.

[1]: https://launchpadlibrarian.net/263258156/buildlog_ubuntu-precise-amd64.lean_0.2.0.20160603023524.gitc73b2860d5211187e9aa1039d1a49dcabdca4292~12.04_BUILDING.txt.gz

[2]: https://cmake.org/cmake/help/v2.8.12/cmake.html#command:set