feat(library/compiler): add compiler specific simplification rules
This just a proof of concept and experiment. This feature is going to be
implemented in Lean4.

@kha Here is the data I managed to collect:

- The `.olean` files did not increase much.
  Before: 6.98Mb
  After:  7.62Mb

- Just running the simplifier introduces a huge overhead even if we
  don't create any auxiliary definition and don't inline anything.
  Here are the numbers for compiling the core lib

  master branch:                           41secs
  simplify only:                      1min 28secs
  simplify + aux definition:          1min 30secs
  simplify + aux definition + inline: 1min 47secs

One big overhead is that the simplifier creates proofs, that will be
deleted later. To do this, it creates big custom congruence lemmas.

It is also missing transformations that are relevant for a compiler.