Register ffi attribute
WIP

We register a declaration attribute to bind foreign functions to Lean
constants like so:

    @[ffi `libfoo]
    constant foo_fun : c_int -> c_int -> c_int

Allowing the vm to get a handle to the foreign function from `dlsym`
and map it with a `ffi_cif` struct from `ffi.h`. This will allow Lean
users to then create functions using this type to marshal Lean values
into the bound `ffi_cif` and make the call with `libffi` under the
hood.